+++ src/anemoi/transform/variables/variables.py
@@ -1,6 +1,8 @@
-# (C) Copyright 2024 European Centre for Medium-Range Weather Forecasts.
+# (C) Copyright 2024 Anemoi contributors.
+#
 # This software is licensed under the terms of the Apache Licence Version 2.0
 # which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
 # In applying this licence, ECMWF does not waive the privileges and immunities
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
@@ -42,8 +44,12 @@ class VariableFromMarsVocabulary(Variable):
         return self.data.get("process") == "accumulation"
 
     @property
+    def is_instantanous(self):
+        return "process" not in self.data.get
+
+    @property
     def grib_keys(self):
-        return self.data.get("mars", {})
+        return self.data.get("mars", {}).copy()
 
 
 class VariableFromDict(VariableFromMarsVocabulary):